pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "posture-recognition"
version = "1.0.0"
description = "PyTorch-based sitting posture recognition with CLI tools"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.26.0,<2.0",
  "opencv-python>=4.10.0",
  "pillow>=10.0.0",
  "scipy>=1.13.0",
  "pandas>=2.2.0",
  "pyyaml>=6.0",
  "click>=8.1,<8.2",
  "typer>=0.12.0",
  "rich>=13.7.0",
  "tqdm>=4.66.0",
  "ultralytics>=8.3.0,<9"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0"
]

[project.scripts]
posture-recognition = "posture_recognition.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q"
testpaths = ["tests"]
